Added a switch to function ApplyWeaponFOV
Added a switch to toggle the ApplyWeaponFOV function on or off due to some weapon mods for UT2004 not playing nice with the way Foxfov changes the weaponFOV.
Ballistic Weapons is 1 of these weapon mods that has some weird effects when the ApplyWeaponFOV is active. Turning off ApplyWeaponFOV restored the guns to there intented positions. 

The switch can be turned on or off via the user.ini and is in the same section where all other settings are stored